Work on parameters

+++ pdf/__init__.py
@@ -24,7 +24,8 @@ def generate_pdf(book_path, base_filename, tex_file):
         os.unlink(pdf_file)
         raise ValueError('Error {} executing command: {}'.format(retcode, ' '.join(cmd)))
     else:
-        os.unlink(filepath(book_path, base_filename, 'toc'))
+        if os.path.isfile(filepath(book_path, base_filename, 'toc')):
+            os.unlink(filepath(book_path, base_filename, 'toc'))
         os.unlink(filepath(book_path, base_filename, 'log'))
         os.unlink(filepath(book_path, base_filename, 'aux'))
         show_file(pdf_file)
